requirements.txt
beautifulsoup4==4.8.2
catboost==0.24.1
empyrical==0.5.3
grequests==0.6.0
hyperopt==0.2.4
hpsklearn==0.1.0
joblib==0.13.2
lightgbm==3.0.0
matplotlib==3.1.0
numpy==1.16.4
optml==0.2.3
pandas==1.0.3
plotly==4.6.0
requests==2.22.0
regex==2020.7.14
shap==0.36.0
scipy==1.4.1
seaborn==0.9.0
scikit-learn==0.23.2
selenium==3.141.0
SQLAlchemy==1.2.19
tqdm==4.47.0
textblob==0.15.3
tensorflow==2.3.0
transformers==2.10.0
xgboost==1.0.2
yfinance==0.1.54

TA_Lib-0.4.18-cp37-cp37m-win_amd64.whl
